+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 - 2.2.0
     - Added Randomized Image Rotator transformer
     - Added One Vs Rest ensemble classifier
+    - Calculate the range of continuous features when describing dataset
+    - Added Gower distance kernel
 
 - 2.1.1
     - Do not consider unset properties when determining revision